Add files via upload
Here is a small repository for the app. All it can do so far is add tasks, mark them as completed, the deletion of completed items, and pinning items as important. It is not yet possible to edit the name of a task, so I will be working on implementing that feature. One thing you will notice right away is that the app is very ugly. I will also be fixing the CSS so that this does not become a persistent problem.

This is a semi-functional build of the app. It has two significant bugs that I need to fix (you are more than welcome to try yourself):
1. Once a task has been pinned, it cannot be unpinned.
2. Sometimes the entire app crashes when you pin a task. No idea what causes this.